fix : rename activities and quick fixes
