+++ auth/app/build.gradle
@@ -74,6 +74,3 @@ dependencies {
 }
 
 apply plugin: 'com.google.gms.google-services'
-repositories {
-    mavenCentral()
-}
