+++ auth/app/src/main/java/com/google/firebase/quickstart/auth/EntryChoiceActivity.kt
@@ -3,8 +3,6 @@ package com.google.firebase.quickstart.auth
 import android.content.Intent
 import com.firebase.example.internal.BaseEntryChoiceActivity
 import com.firebase.example.internal.Choice
-import com.google.firebase.quickstart.auth.java.ChooserActivity
-import com.google.firebase.quickstart.auth.kotlin.KotlinChooserActivity
 
 class EntryChoiceActivity : BaseEntryChoiceActivity() {
 
@@ -17,7 +15,7 @@ class EntryChoiceActivity : BaseEntryChoiceActivity() {
                 Choice(
                         "Kotlin",
                         "Run the Firebase Auth quickstart written in Kotlin.",
-                        Intent(this, KotlinChooserActivity::class.java))
+                        Intent(this, ChooserActivity::class.java))
         )
     }
